split reviwews and fixes

+++ src/main/java/com/MovieReviews/moviereviews/repositories/FilmRepository.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface FilmRepository extends JpaRepository<Film, Long> {
+public interface FilmRepository extends JpaRepository<Film, Integer> {
     Optional<Object> findByTitle(String title);
 }
